pyproject.toml
[tool.poetry]
name = "kafka-performance"
version = "0.1.0"
description = ""
authors = ["c2io"]

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.5.3"
locust = "^1.5.3"
arrow = "^1.1.0"
kafka-python = "^2.0.2"
typer = "^0.3.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^21.6b0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
